added better error handling and better jwt validation

+++ resources/address.py
@@ -1,5 +1,5 @@
-import uuid
-from flask import request, session
+from flask_jwt_extended import jwt_required, get_jwt
+from flask import abort, request, session
 from flask.views import MethodView
 from flask_smorest import Blueprint
 from db import db
@@ -13,6 +13,7 @@ from schemas import AddressSchema
 blp = Blueprint("addresses", __name__, description="Operations on addresses")
 
 
+@jwt_required()
 @blp.route("/addresses")
 class Address(MethodView):
     @blp.response(200, AddressSchema(many=True))
@@ -20,6 +21,7 @@ class Address(MethodView):
         return AddressModel.query.all()
 
 
+@jwt_required()
 @blp.route("/address/<string:id>")
 class Address(MethodView):
     @blp.response(200, AddressSchema)
@@ -28,13 +30,16 @@ class Address(MethodView):
         return address
 
     def delete(self, id):
+        jwt = get_jwt()
+        if not jwt.get("is_admin"):
+            abort(401, message="Admin privilege required.")
         address = AddressModel.query.get(id)
         db.session.delete(address)
         db.session.commit()
         return "address deleted"
 
 
-
+@jwt_required()
 @blp.route("/address")
 class Address(MethodView):
     @blp.arguments(AddressSchema)
